Added orderform_id attribute to migration, model, validations and controller. Additionally tweaked tests and documentation to include the property.
I also discovered that the controller was using a method of adding properties to the model that was potentially unsafe. It allowed for non-validated properties to be updated in the patch request which wasn't good practice, so did a minor refactor to remove that method, and implement the $request->only as we do elsewhere in the application.

#1016 add orderform property to discount plans