Tweaked management network creation to include the correct generation of the /28 subnet address for the network.
Added in additional code to make sure that the subnet is unique within the az, by getting a collection and iterating through it to make sure that the subnet does not reside within a used subnet range (cannot rely simply on string comparison on the subnet value).

Adjusted the unit test to incorporate a section to test the subnet calculation.

#1078 eCloud VPC | LBv2 - management networking